requirements.txt
WTForms==2.3.3
Flask_WTF==0.14.3
SQLAlchemy==1.3.22
Flask_SQLAlchemy==2.4.4
alembic==1.4.3
Flask_Migrate==2.5.3
Flask==1.1.2
python-dotenv==0.15.0
